[Refactoring] Compiler warnings removed.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@3482 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/trunk/vnc_winsrc/wsconfig/ConfigDialog.cpp
@@ -50,7 +50,7 @@ void ConfigDialog::initControls()
 void ConfigDialog::loadSettings()
 {
   m_config.loadFromStorage(m_settingsManager);
-  for (int i = 0; i < m_config.getPortMappingContainer().size(); i++) {
+  for (size_t i = 0; i < m_config.getPortMappingContainer().size(); i++) {
     m_ctrlMappingListBox.insertString(i, (TCHAR *)m_config.getPortMappingContainer().at(i).toString().c_str());
   }
 }
